Added working spline fit example, merged operator examples in single file.
